# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface IpcSection {
  /** e.g. 'IPC 420' */
  section: string;
  title: string;
  description: string;
  /** 0..1 */
  confidence: number;
  /** Keywords or phrases from the incident that triggered this section */
  triggered_by: string[];
}

export interface ConstitutionArticle {
  /** e.g. 'Article 21' */
  article: string;
  title: string;
  description: string;
  confidence: number;
}

export type EvidenceItemType =
  (typeof EvidenceItemType)[keyof typeof EvidenceItemType];

export const EvidenceItemType = {
  document: "document",
  witness: "witness",
  digital: "digital",
  physical: "physical",
  medical: "medical",
  other: "other",
} as const;

export type EvidenceItemPriority =
  (typeof EvidenceItemPriority)[keyof typeof EvidenceItemPriority];

export const EvidenceItemPriority = {
  critical: "critical",
  important: "important",
  optional: "optional",
} as const;

export interface EvidenceItem {
  id: string;
  type: EvidenceItemType;
  description: string;
  priority: EvidenceItemPriority;
  collected: boolean;
}

export interface PrecedentCase {
  title: string;
  court: string;
  year: number;
  summary: string;
  outcome: string;
  similarity: number;
}

export interface RightsViolation {
  /** e.g. 'Right to Life (Article 21)' */
  right: string;
  description: string;
  complaint_channel: string;
}

export interface StrengthScore {
  /** 0..100 */
  overall: number;
  evidence: number;
  legal_coverage: number;
  precedent: number;
  /** e.g. 'Strong', 'Moderate', 'Weak' */
  verdict: string;
  notes: string;
}

export type FirDraftTimelineItem = {
  when: string;
  event: string;
};

export interface FirDraft {
  complainant_name: string;
  complainant_address: string;
  incident_date: string;
  incident_location: string;
  incident_summary: string;
  sections_invoked: string[];
  timeline: FirDraftTimelineItem[];
  accused_details: string;
  prayer: string;
}

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  draft: "draft",
  analyzed: "analyzed",
  fir_generated: "fir_generated",
  filed: "filed",
} as const;

export interface Case {
  id: string;
  title: string;
  complainant_name: string;
  incident_text: string;
  input_language: string;
  /** e.g. 'Cyber Fraud', 'Assault', 'Domestic Violence' */
  category: string;
  status: CaseStatus;
  summary: string;
  /** Plain-language explanation of the law that applies */
  explanation: string;
  ipc_sections: IpcSection[];
  constitution_articles: ConstitutionArticle[];
  evidence: EvidenceItem[];
  precedents: PrecedentCase[];
  rights_violations: RightsViolation[];
  strength: StrengthScore;
  fir_draft?: FirDraft | null;
  /** XAI reasoning trace */
  reasoning: string;
  created_at: string;
  updated_at: string;
}

export type NewCaseInputInputLanguage =
  (typeof NewCaseInputInputLanguage)[keyof typeof NewCaseInputInputLanguage];

export const NewCaseInputInputLanguage = {
  en: "en",
  hi: "hi",
  te: "te",
} as const;

export interface NewCaseInput {
  complainant_name: string;
  incident_text: string;
  input_language: NewCaseInputInputLanguage;
}

export interface EvidenceUpdate {
  evidence_id: string;
  collected: boolean;
}

export type TranslateRequestTargetLanguage =
  (typeof TranslateRequestTargetLanguage)[keyof typeof TranslateRequestTargetLanguage];

export const TranslateRequestTargetLanguage = {
  en: "en",
  hi: "hi",
  te: "te",
} as const;

export interface TranslateRequest {
  target_language: TranslateRequestTargetLanguage;
}

export interface TranslationResponse {
  target_language: string;
  summary: string;
  explanation: string;
}

export interface TranscribeRequest {
  audio_base64: string;
  mime_type: string;
  language?: string;
}

export interface TranscriptionResponse {
  text: string;
  detected_language: string;
}

export type DashboardSummaryByCategoryItem = {
  category: string;
  count: number;
};

export interface DashboardSummary {
  total_cases: number;
  firs_generated: number;
  evidence_collected: number;
  evidence_pending: number;
  avg_strength: number;
  by_category: DashboardSummaryByCategoryItem[];
}

export interface ActivityItem {
  case_id: string;
  case_title: string;
  action: string;
  at: string;
}

export interface LawFrequency {
  section: string;
  title: string;
  count: number;
}

export type DeleteCase200 = {
  success: boolean;
};