# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from "./caseStatus";
import type { ConstitutionArticle } from "./constitutionArticle";
import type { EvidenceItem } from "./evidenceItem";
import type { FirDraft } from "./firDraft";
import type { IpcSection } from "./ipcSection";
import type { PrecedentCase } from "./precedentCase";
import type { RightsViolation } from "./rightsViolation";
import type { StrengthScore } from "./strengthScore";

export interface Case {
  id: string;
  title: string;
  complainant_name: string;
  incident_text: string;
  input_language: string;
  /** e.g. 'Cyber Fraud', 'Assault', 'Domestic Violence' */
  category: string;
  status: CaseStatus;
  summary: string;
  /** Plain-language explanation of the law that applies */
  explanation: string;
  ipc_sections: IpcSection[];
  constitution_articles: ConstitutionArticle[];
  evidence: EvidenceItem[];
  precedents: PrecedentCase[];
  rights_violations: RightsViolation[];
  strength: StrengthScore;
  fir_draft?: FirDraft | null;
  /** XAI reasoning trace */
  reasoning: string;
  created_at: string;
  updated_at: string;
}