# lib/api-zod/src/generated/types/constitutionArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ConstitutionArticle {
  /** e.g. 'Article 21' */
  article: string;
  title: string;
  description: string;
  confidence: number;
}