# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryByCategoryItem } from "./dashboardSummaryByCategoryItem";

export interface DashboardSummary {
  total_cases: number;
  firs_generated: number;
  evidence_collected: number;
  evidence_pending: number;
  avg_strength: number;
  by_category: DashboardSummaryByCategoryItem[];
}