# lib/api-zod/src/generated/types/evidenceItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceItemPriority } from "./evidenceItemPriority";
import type { EvidenceItemType } from "./evidenceItemType";

export interface EvidenceItem {
  id: string;
  type: EvidenceItemType;
  description: string;
  priority: EvidenceItemPriority;
  collected: boolean;
}