# lib/api-zod/src/generated/types/ipcSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface IpcSection {
  /** e.g. 'IPC 420' */
  section: string;
  title: string;
  description: string;
  /** 0..1 */
  confidence: number;
  /** Keywords or phrases from the incident that triggered this section */
  triggered_by: string[];
}