# lib/api-zod/src/generated/types/newCaseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { NewCaseInputInputLanguage } from "./newCaseInputInputLanguage";

export interface NewCaseInput {
  complainant_name: string;
  incident_text: string;
  input_language: NewCaseInputInputLanguage;
}