# lib/api-zod/src/generated/types/strengthScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface StrengthScore {
  /** 0..100 */
  overall: number;
  evidence: number;
  legal_coverage: number;
  precedent: number;
  /** e.g. 'Strong', 'Moderate', 'Weak' */
  verdict: string;
  notes: string;
}