# lib/api-zod/src/generated/types/transcribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeRequest {
  audio_base64: string;
  mime_type: string;
  language?: string;
}