# lib/api-zod/src/generated/types/transcriptionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptionResponse {
  text: string;
  detected_language: string;
}