# lib/api-zod/src/generated/types/translateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nyaya AI Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { TranslateRequestTargetLanguage } from "./translateRequestTargetLanguage";

export interface TranslateRequest {
  target_language: TranslateRequestTargetLanguage;
}